Move log and settings into security module
Move the `SecurityLog` and `SecuritySettings` into separate sub-packages
of the enterprise security module.
