+++ community/security/src/main/java/org/neo4j/server/security/auth/AuthProcedures.java
@@ -51,7 +51,6 @@ public class AuthProcedures
             throws InvalidArgumentsException, IOException
     {
         BasicAuthSubject subject = BasicAuthSubject.castOrFail( authSubject );
-//        subject.getAuthManager().newUser( username, password, requirePasswordChange );
         userManager.newUser( username, password, requirePasswordChange );
     }
