+++ manual/config-docs/pom.xml
@@ -123,7 +123,7 @@
                   <arg value="org.neo4j.kernel.ha.HaSettings" />
                   <arg value="org.neo4j.consistency.ConsistencyCheckSettings" />
                   <arg value="org.neo4j.server.enterprise.EnterpriseServerSettings" />
-                  <arg value="org.neo4j.server.security.enterprise.auth.SecuritySettings" />
+                  <arg value="org.neo4j.server.security.enterprise.configuration.SecuritySettings" />
                   <arg value="org.neo4j.kernel.impl.enterprise.configuration.EnterpriseEditionSettings" />
                 </java>
